% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ags-package.R
\docType{package}
\name{ags-package}
\alias{ags-package}
\alias{ags}
\title{Tools to work with the German AGS.}
\description{
Useful tools to work with German AGS.
}
\details{
Feedback is very welcome!


\tabular{ll}{
	Package: \tab ags \cr
	Type: \tab Package\cr
	Version: \tab 0.0.0.9000\cr
	Date: \tab 2016-11-20\cr
	License: \tab  GPL-3\cr
	}
}
\author{
Moritz Marbach \email{marbach@ipz.uzh.ch}
}
