% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ags_dist.R
\name{ags_dist}
\alias{ags_dist}
\title{Calculates the distance between two AGS numbers}
\usage{
ags_dist(x, y, landw = 10^6, kreisw = 10^3, gemw = 1,
  ceiling = 99999999)
}
\description{
Calculates the distance between two AGS numbers
}
\details{
The default is just the difference between 
the two integer numbers.
}
