% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw_sn.R
\docType{data}
\name{btw_sn}
\alias{btw_sn}
\title{Number of voters and valid votes in Saxony (1994-2017)}
\format{
A data frame with 155 rows and 4 variables:
\describe{
\item{district}{AGS of the district.}
\item{year}{Election year.}
\item{voters}{Number of eligible voters.}
\item{valid}{Number of valid votes.}
}
}
\source{
\url{https://www.regionalstatistik.de}
}
\usage{
btw_sn
}
\description{
The dataset includes the number of voters and valid votes
in all federal elections (Bundestagswahlen) across
districts in Saxony.
}
\keyword{datasets}
